+++ webdev/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.7.8-dev
 
 - Update `vm_service` to `^8.1.0`.
+- Enable expression evaluation by default.
 
 ## 2.7.7